feat: add Home page and redo app structure
