+++ src/main.tsx
@@ -1,17 +1,5 @@
-import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import { BrowserRouter, Routes, Route } from "react-router";
 import "./index.css";
-import App from "./App.tsx";
-import { Documentation } from "./pages/Documentation/Documentation.tsx";
+import { App } from "./App.tsx";
 
-createRoot(document.getElementById("root")!).render(
-  <StrictMode>
-    <BrowserRouter>
-      <Routes>
-        <Route path="/" element={<App />} />
-        <Route path="documentation" element={<Documentation />} />
-      </Routes>
-    </BrowserRouter>
-  </StrictMode>
-);
+createRoot(document.getElementById("root")!).render(<App />);
