+++ src/pages/Home/Home.css
@@ -1,0 +1,11 @@
+body {
+  margin: 0;
+}
+
+.tree-container {
+  width: 100%;
+  flex: 1;
+  margin: 0 auto;
+  box-sizing: border-box;
+  overflow: hidden;
+}
